Move print and share methods out and clean up used pages
