+++ src/scss/boards.scss
@@ -184,19 +184,6 @@
   }
 }
 
-.ru_share {
-  background-image: url(/assets/share.svg);
-  width: 50px;
-  height: 58px;
-  background-repeat: no-repeat;
-  z-index: 1;
-  float: right;
-  position: relative;
-  margin-right: 10px;
-  margin-left: 10px;
-  background-color: transparent;
-}
-
 .share-link-table {
   overflow: hidden;
   white-space: nowrap;
